requirements.txt
confluent-kafka==2.3.0
pydantic==2.7.0
neo4j==5.18.0
elasticsearch==8.13.0
fastapi==0.110.0
uvicorn==0.29.0
langchain==0.1.13
langchain-openai==0.1.0
instructor==1.1.0
pdfplumber==0.11.0
pdf2image==1.17.0
pytesseract==0.3.10
httpx==0.27.0
redis==5.0.3
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-instrumentation-fastapi==0.44b0
langfuse==2.21.0
pytest==8.1.1
pytest-asyncio==0.23.6
